Add tests for chapter 6
For test 6.1.11 & 6.1.12 the remediation could not be tried successfully. The test on the other hand does what it is supposed to do. Therefore we will see in the later implementation for the ecosystem if we can find a solution or if it is just a local problem with my vagrant machine.